fail to add offense rating effect on chances, but adding the 'no score' abilities
before this feature the team would always score. Now there is a chance that the team will
not score any points on a given possession. Need to work on how to get different values from
different instances as a general value, i.e. workout answer to a.value - b.value. Will ask
tomorrow